+++ CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 <!-- ## [Unreleased] -->
 
+## [0.3.3] - 2022-03-25
+
+### Fixed
+
+- correctly display multi-(c)hunk diffs
+
+### Removed
+
+- verbose option no longer affects diff print
+
 ## [0.3.2] - 2022-03-24
 
 ### Added